Cross-gh/rm signals
